+++ service/cziti/dns.go
@@ -43,7 +43,7 @@ var DNS = &dnsImpl{}
 
 type dnsImpl struct {
 	cidr    uint32
-	counter uint32
+	ipCount uint32
 
 	serviceMap map[string]ctxService
 
@@ -67,12 +67,13 @@ type ctxService struct {
 }
 
 func normalizeDnsName(dnsName string) string {
-	if strings.HasSuffix(dnsName, ".") {
-		return dnsName
-	} else {
+	dnsName = strings.TrimSpace(dnsName)
+	if !strings.HasSuffix(dnsName, ".") {
 		// append a period to the dnsName - forcing it to be a FQDN
-		return dnsName + "."
+		dnsName += "."
 	}
+
+	return strings.ToLower(dnsName) //normalize the casing
 }
 
 // RegisterService will return the next ip address in the configured range. If the ip address is not
@@ -115,7 +116,7 @@ func (dns *dnsImpl) RegisterService(svcId string, dnsNameToReg string, port uint
 		}
 	} else {
 		// if not used at all - map it
-		nextAddr := dns.cidr | atomic.AddUint32(&dns.counter, 1)
+		nextAddr := dns.cidr | atomic.AddUint32(&dns.ipCount, 1)
 		ip = make(net.IP, 4)
 		binary.BigEndian.PutUint32(ip, nextAddr)
 
@@ -142,19 +143,25 @@ func (dns *dnsImpl) Resolve(toResolve string) net.IP {
 }
 
 func (dns *dnsImpl) DeregisterService(ctx *CZitiCtx, name string) {
+	log.Warnf("DEREG SERVICE: before for loop")
 	for k, sc := range dns.serviceMap {
+		log.Warnf("DEREG SERVICE: iterating sc.ctx:%v, sc.name:%v", sc.ctx, sc.name)
 		if sc.ctx == ctx && sc.name == name {
+			log.Warnf("DEREG SERVICE: if was true")
 			if sc.count < 1 {
-				log.Infof("removing %s from DNS mapping", name)
+				log.Infof("removing service named %s from DNS mapping known as %s", name, k)
 				delete(dns.serviceMap, k)
 			} else {
 				// another service is using the mapping - can't remove it yet so decrement
-				sc.count --
 				log.Debugf("cannot remove dns mapping for %s yet - %d other services still use this hostname", name, sc.count)
+				sc.count --
 			}
 			return
+		} else {
+			log.Warnf("DEREG SERVICE: if was false")
 		}
 	}
+	log.Warnf("DEREG SERVICE: for loop completed")
 }
 
 func (this *dnsImpl) GetService(ip net.IP, port uint16) (*CZitiCtx, string, error) {
@@ -182,6 +189,6 @@ func DnsInit(ip string, maskBits int) {
 		i := net.ParseIP(ip).To4()
 		mask := net.CIDRMask(maskBits, 32)
 		DNS.cidr = binary.BigEndian.Uint32(i) & binary.BigEndian.Uint32(mask)
-		DNS.counter = 2
+		DNS.ipCount = 2
 	})
 }